requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
datasets
transformers
